fix: internal conflict declaration
Arch community recommends that '*-git' conflicts with the original one, not the other way around
(also, not both ways).

For the '*-all' package, we don't need to specify a conflict with each theme package. Instead, we
specify a conflict with any 'tela-circle-icon-theme' provider, which includes every theme package.

Additionally, we don't specify a conflict with '*-all' for each theme package, because the
conflict is already specified in the '*-all' package.