fix: internal conflict declaration
We don't need to specify that '*-git' package conflicts with themselves, just its non git version.

For the '*-all' package, we don't need to specify a conflict with each theme package. Instead, we
specify a conflict with any 'tela-circle-icon-theme' provider, which includes every theme package.

Additionally, we don't specify a conflict with '*-all' for each theme package, because the
conflict is already specified in the '*-all' package.